"Push Swap tested"

+++ libft/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: asolano- <asolano-@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/14 12:28:40 by asolano-          #+#    #+#             */
-/*   Updated: 2022/06/15 08:55:32 by asolano-         ###   ########.fr       */
+/*   Updated: 2022/06/16 12:15:02 by asolano-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_abs(int num)
+int	ft_abs(float num)
 {
 	int	tmp;
